refactor(kamn-e2e-harness): extract run-contract execution module (#5688)
